requirements.txt
# Core ML and Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# MLflow for experiment tracking and model registry
mlflow>=2.8.0

# FastAPI for API deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data ingestion
kagglehub>=0.1.0
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.25.0 